+++ readium/lcp/r2-lcp/src/main/java/org/readium/r2/lcp/license/model/components/lsd/PotentialRights.kt
@@ -9,13 +9,11 @@
 
 package org.readium.r2.lcp.license.model.components.lsd
 
-import org.joda.time.DateTime
 import org.json.JSONObject
+import org.readium.r2.shared.extensions.iso8601ToDate
+import org.readium.r2.shared.extensions.optNullableString
+import java.util.*
 
 data class PotentialRights(val json: JSONObject) {
-    val end: DateTime?
-
-    init {
-        end = if (json.has("end")) DateTime(json.getString("end")) else null
-    }
+    val end: Date? = json.optNullableString("end")?.iso8601ToDate()
 }